+++ components/dashboard/src/teams/TeamSettings.tsx
@@ -5,13 +5,13 @@
  */
 
 import { OrganizationSettings } from "@gitpod/public-api/lib/gitpod/v1/organization_pb";
-import React, { Children, FormEvent, ReactNode, useCallback, useMemo, useState } from "react";
+import React, { Children, ReactNode, useCallback, useMemo, useState } from "react";
 import Alert from "../components/Alert";
 import ConfirmationModal from "../components/ConfirmationModal";
 import { InputWithCopy } from "../components/InputWithCopy";
 import Modal, { ModalBody, ModalFooter, ModalHeader } from "../components/Modal";
 import { InputField } from "../components/forms/InputField";
-import { TextInput, TextInputField } from "../components/forms/TextInputField";
+import { TextInputField } from "../components/forms/TextInputField";
 import { Heading2, Heading3, Subheading } from "../components/typography/headings";
 import { useIsOwner } from "../data/organizations/members-query";
 import { useOrgSettingsQuery } from "../data/organizations/org-settings-query";
@@ -20,7 +20,7 @@ import { useUpdateOrgMutation } from "../data/organizations/update-org-mutation"
 import { useUpdateOrgSettingsMutation } from "../data/organizations/update-org-settings-mutation";
 import { useOnBlurError } from "../hooks/use-onblur-error";
 import { ReactComponent as Stack } from "../icons/Stack.svg";
-import { converter, organizationClient } from "../service/public-api";
+import { organizationClient } from "../service/public-api";
 import { gitpodHostUrl } from "../service/service";
 import { useCurrentUser } from "../user-context";
 import { OrgSettingsPage } from "./OrgSettingsPage";
@@ -30,13 +30,8 @@ import { useInstallationDefaultWorkspaceImageQuery } from "../data/installation/
 import { ConfigurationSettingsField } from "../repositories/detail/ConfigurationSettingsField";
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@podkit/select/Select";
 import { useDocumentTitle } from "../hooks/use-document-title";
-import { LoadingButton } from "@podkit/buttons/LoadingButton";
 import { PlainMessage } from "@bufbuild/protobuf";
-import { CheckboxInputField } from "../components/forms/CheckboxInputField";
-import { WorkspaceTimeoutDuration } from "@gitpod/gitpod-protocol";
 import { useToast } from "../components/toasts/Toasts";
-import { Link } from "react-router-dom";
-import { useOrgBillingMode } from "../data/billing-mode/org-billing-mode-query";
 
 export default function TeamSettingsPage() {
     useDocumentTitle("Organization Settings - General");
@@ -50,11 +45,6 @@ export default function TeamSettingsPage() {
     const [teamNameToDelete, setTeamNameToDelete] = useState("");
     const [teamName, setTeamName] = useState(org?.name || "");
     const [updated, setUpdated] = useState(false);
-
-    const billingMode = useOrgBillingMode();
-    const [workspaceTimeout, setWorkspaceTimeout] = useState<string | undefined>(undefined);
-    const [allowTimeoutChangeByMembers, setAllowTimeoutChangeByMembers] = useState<boolean | undefined>(undefined);
-    const [workspaceTimeoutSettingError, setWorkspaceTimeoutSettingError] = useState<string | undefined>(undefined);
 
     const updateOrg = useUpdateOrgMutation();
 
@@ -107,15 +97,6 @@ export default function TeamSettingsPage() {
 
     const [showImageEditModal, setShowImageEditModal] = useState(false);
 
-    useMemo(() => {
-        setWorkspaceTimeout(
-            settings?.timeoutSettings?.inactivity
-                ? converter.toDurationString(settings.timeoutSettings.inactivity)
-                : undefined,
-        );
-        setAllowTimeoutChangeByMembers(!settings?.timeoutSettings?.denyUserTimeouts);
-    }, [settings?.timeoutSettings]);
-
     const handleUpdateTeamSettings = useCallback(
         async (newSettings: Partial<PlainMessage<OrganizationSettings>>, options?: { throwMutateError?: boolean }) => {
             if (!org?.id) {
@@ -129,7 +110,6 @@ export default function TeamSettingsPage() {
                     ...settings,
                     ...newSettings,
                 });
-                setWorkspaceTimeoutSettingError(undefined);
                 toast("Organization settings updated");
             } catch (error) {
                 if (options?.throwMutateError) {
@@ -141,36 +121,6 @@ export default function TeamSettingsPage() {
         },
         [updateTeamSettings, org?.id, isOwner, settings, toast],
     );
-
-    const handleUpdateOrganizationTimeoutSettings = useCallback(
-        (e: FormEvent<HTMLFormElement>) => {
-            e.preventDefault();
-            try {
-                if (workspaceTimeout) {
-                    WorkspaceTimeoutDuration.validate(workspaceTimeout);
-                }
-            } catch (error) {
-                setWorkspaceTimeoutSettingError(error.message);
-                return;
-            }
-
-            // Nothing has changed
-            if (workspaceTimeout === undefined && allowTimeoutChangeByMembers === undefined) {
-                return;
-            }
-
-            handleUpdateTeamSettings({
-                timeoutSettings: {
-                    inactivity: workspaceTimeout ? converter.toDuration(workspaceTimeout) : undefined,
-                    denyUserTimeouts: !allowTimeoutChangeByMembers,
-                },
-            });
-        },
-        [workspaceTimeout, allowTimeoutChangeByMembers, handleUpdateTeamSettings],
-    );
-
-    const billingModeAllowsWorkspaceTimeouts =
-        billingMode.data?.mode === "none" || (billingMode.data?.mode === "usage-based" && billingMode.data?.paid);
 
     return (
         <>
@@ -255,60 +205,6 @@ export default function TeamSettingsPage() {
                             installationDefaultWorkspaceImage={installationDefaultImage}
                             onClick={() => setShowImageEditModal(true)}
                         />
-                    </ConfigurationSettingsField>
-
-                    <ConfigurationSettingsField>
-                        <Heading3>Workspace timeouts</Heading3>
-                        {!billingModeAllowsWorkspaceTimeouts && (
-                            <Alert type="info" className="my-3">
-                                Setting Workspace timeouts is only available for organizations on a paid plan. Visit{" "}
-                                <Link to={"/billing"} className="gp-link">
-                                    Billing
-                                </Link>{" "}
-                                to upgrade your plan.
-                            </Alert>
-                        )}
-                        <form onSubmit={handleUpdateOrganizationTimeoutSettings}>
-                            <InputField
-                                label="Default workspace timeout"
-                                error={workspaceTimeoutSettingError}
-                                hint={
-                                    <span>
-                                        Use minutes or hours, like <span className="font-semibold">30m</span> or{" "}
-                                        <span className="font-semibold">2h</span>
-                                    </span>
-                                }
-                            >
-                                <TextInput
-                                    value={workspaceTimeout ?? ""}
-                                    placeholder="e.g. 30m"
-                                    onChange={setWorkspaceTimeout}
-                                    disabled={
-                                        updateTeamSettings.isLoading || !isOwner || !billingModeAllowsWorkspaceTimeouts
-                                    }
-                                />
-                            </InputField>
-                            <CheckboxInputField
-                                label="Allow members to change workspace timeouts"
-                                hint="Allow users to change the timeout duration for their workspaces as well as setting a default one in their user settings."
-                                checked={!!allowTimeoutChangeByMembers}
-                                containerClassName="my-4"
-                                onChange={setAllowTimeoutChangeByMembers}
-                                disabled={
-                                    updateTeamSettings.isLoading || !isOwner || !billingModeAllowsWorkspaceTimeouts
-                                }
-                            />
-                            <LoadingButton
-                                type="submit"
-                                loading={updateTeamSettings.isLoading}
-                                disabled={
-                                    workspaceTimeout ===
-                                        converter.toDurationString(user?.workspaceTimeoutSettings?.inactivity) ?? ""
-                                }
-                            >
-                                Save
-                            </LoadingButton>
-                        </form>
                     </ConfigurationSettingsField>
 
                     {showImageEditModal && (
